requirements.txt
google-adk>=1.21.0,<2.0.0
google-genai>=1.31.0,<2.0.0
mcp>=1.0.0,<2.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
httpx>=0.28.0,<1.0.0
